feat: for production, deploy API and SPA separately
Azure App Service free tier does not support custom domains, so we're running the SPA as a Static Web App and keeping the API in the App Service.
This requires that we don't build the SPA when building the solution as the SPA will be built and deployed seperately.
For local development, it should continue to work as before.